feat(pci.private-network): implementing editable allocation pool
ref: TAPC-1860
