+++ packages/manager/apps/pci-private-network/src/pages/new/subnet/SubnetConfig.spec.tsx
@@ -29,6 +29,7 @@ describe('SubnetConfig CIDR', () => {
         const form = useForm<NewPrivateNetworkForm>({
           defaultValues: {
             subnet: {
+              allocationPools: [],
               cidr,
             },
           },
